Finish migration to JDK7
Multicatch, for-loop, switch over strings, try catch, etc.


Former-commit-id: 53c5f9841a6bfdbdd582fb9b21420579a283ce31

+++ seqware-portal/src/main/java/net/sourceforge/solexatools/webapp/controller/StudyReportBoxController.java
@@ -172,7 +172,7 @@ public class StudyReportBoxController extends BaseCommandController {
                     for (Sample sample : childSamples) {
                         Sample rootSample = sampleService.getRootSample(sample);
                         sb.append(rootSample.getTitle() + "\t");
-                        sb.append(rootSample.getSampleId().intValue() != sample.getSampleId().intValue() ? sample.getTitle() + "\t"
+                        sb.append(rootSample.getSampleId() != sample.getSampleId() ? sample.getTitle() + "\t"
                                 : "no child" + "\t");
                         for (Workflow workflow : workflows) {
                             Status status = sampleReportService.getStatus(currentStudy, sample, workflow);
